+++ packages/jest-cli/package.json
@@ -5,6 +5,7 @@
   "main": "build/jest.js",
   "dependencies": {
     "@jest/reporters": "^24.1.0",
+    "@jest/transform": "^24.1.0",
     "ansi-escapes": "^3.0.0",
     "chalk": "^2.0.1",
     "exit": "^0.1.2",
