+++ packages/jest-reporters/package.json
@@ -4,10 +4,10 @@
   "version": "24.1.0",
   "main": "build/index.js",
   "dependencies": {
+    "@jest/transform": "^24.1.0",
     "chalk": "^2.0.1",
     "exit": "^0.1.2",
     "glob": "^7.1.2",
-    "jest-runtime": "^24.1.0",
     "istanbul-api": "^2.1.1",
     "istanbul-lib-coverage": "^2.0.2",
     "istanbul-lib-instrument": "^3.0.1",
