+++ packages/jest-runtime/package.json
@@ -9,12 +9,9 @@
   "license": "MIT",
   "main": "build/index.js",
   "dependencies": {
-    "@babel/core": "^7.1.0",
-    "babel-plugin-istanbul": "^5.1.0",
+    "@jest/transform": "^24.1.0",
     "chalk": "^2.0.1",
-    "convert-source-map": "^1.4.0",
     "exit": "^0.1.2",
-    "fast-json-stable-stringify": "^2.0.0",
     "glob": "^7.1.3",
     "graceful-fs": "^4.1.15",
     "jest-config": "^24.1.0",
@@ -25,24 +22,18 @@
     "jest-snapshot": "^24.1.0",
     "jest-util": "^24.0.0",
     "jest-validate": "^24.0.0",
-    "micromatch": "^3.1.10",
     "realpath-native": "^1.1.0",
     "slash": "^2.0.0",
     "strip-bom": "^3.0.0",
-    "write-file-atomic": "2.4.1",
     "yargs": "^12.0.2"
   },
   "devDependencies": {
-    "@types/babel__core": "^7.0.4",
-    "@types/convert-source-map": "^1.5.1",
     "@types/exit": "^0.1.30",
     "@types/glob": "^7.1.1",
     "@types/graceful-fs": "^4.1.2",
-    "@types/micromatch": "^3.1.0",
-    "@types/strip-bom": "3.0.0",
-    "@types/write-file-atomic": "^2.1.1",
+    "@types/slash": "^2.0.0",
+    "@types/strip-bom": "^3.0.0",
     "@types/yargs": "^12.0.2",
-    "jest-environment-jsdom": "^24.0.0",
     "jest-environment-node": "^24.0.0"
   },
   "bin": {
