+++ packages/jest-runtime/src/index.js
@@ -24,10 +24,9 @@ import Resolver from 'jest-resolve';
 import {createDirectory, deepCyclicCopy} from 'jest-util';
 import {escapePathForRegex} from 'jest-regex-util';
 import Snapshot from 'jest-snapshot';
+import {ScriptTransformer, shouldInstrument} from '@jest/transform';
 import fs from 'graceful-fs';
 import stripBOM from 'strip-bom';
-import ScriptTransformer from './ScriptTransformer';
-import shouldInstrument from './shouldInstrument';
 import {run as cliRun} from './cli';
 import {options as cliOptions} from './cli/args';
 import {findSiblingsWithFileExtension} from './helpers';
