+++ packages/jest-transform/src/index.js
@@ -1,0 +1,11 @@
+/**
+ * Copyright (c) Facebook, Inc. and its affiliates. All Rights Reserved.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ *
+ * @flow
+ */
+
+export {default as ScriptTransformer} from './ScriptTransformer';
+export {default as shouldInstrument} from './shouldInstrument';
